requirements.txt
streamlit==1.26.0
diffusers==0.18.0
torch==2.2.2
transformers==4.35.0
Pillow==9.5.0
accelerate==0.23.0
scipy==1.10.1
Flask==2.1.2

